+++ apps/desktop/src/components/pane-shell/tree/renderer/tree-group.tsx
@@ -472,12 +472,8 @@ export function TreeGroup({
                     role="tab"
                     style={{ cursor: 'grab' }}
                   >
-                    {chrome.accent ? (
-                      <span
-                        aria-hidden="true"
-                        className="ml-2 -mr-1 size-1 shrink-0 rounded-full"
-                        style={{ backgroundColor: chrome.accent }}
-                      />
+                    {chrome.tabLead ? (
+                      <span className="ml-2 -mr-1 flex shrink-0 items-center">{chrome.tabLead()}</span>
                     ) : null}
                     <PaneTabLabel>{title}</PaneTabLabel>
                   </PaneTab>
